requirements.txt
# Emare Code Dependencies

# Core
python-dotenv>=1.0.0
pydantic>=2.0.0

# AI Providers
google-genai>=0.3.0
openai>=1.0.0

# Database
sqlalchemy>=2.0.0

# CLI (optional)
rich>=13.0.0

# Testing (optional)
pytest>=7.0.0
